requirements.txt
newsapi-python==0.2.7
nltk==3.8.1
pandas==2.1.4
